==> 3_FeatureSelect.ipynb <==
# Feature Selection

In this notebook features are ranked using importance and 2 sets 20- and 40- most important features are determined

import pandas as pd
import numpy as np

df = pd.read_csv("data/train_data.tsv.gz", sep="\t", compression="gzip")
df = df.drop(columns=['Unnamed: 0'])
df.head()

## XGBoost Feature Importance

Use aggregate score as y

from xgboost import XGBRegressor

X = df.iloc[:,0:399]
y = df.iloc[:,-1]

model = XGBRegressor(objective='reg:logistic')
model.fit(X,y)

importance = model.feature_importances_
importance = pd.DataFrame(importance, index=X.columns, 
                          columns=["Importance"])

## Export 
feature_importance_list.xlsx contains a reviewed version of results.    
20 and 40 most-important features are picked.

importance.sort_values(['Importance'], ascending=0).to_csv('data/feature_importance_results5.csv',sep='\t')

## One Hot Encoding 

1. Remove the columns so that we have our test and train datasets containing top 20 and top 40 most important features.

2. Apply one hot encoding to categorical features for both 20 and 40 feature sets.

df_train_imp0 = pd.read_csv("data/train_data.tsv.gz", sep="\t", compression="gzip")
df_test_imp0 = pd.read_csv("data/test_data.tsv.gz", sep="\t", compression="gzip")

# get categorical features as a list
df_features40 = pd.read_csv('data/feature_importance_results5.csv',sep='\t')[:40]
categorical_features = []
for i in range(len(df_features40['Categorical'])):
    if df_features40['Categorical'][i] != 0:
        categorical_features.append(df_features40['Unnamed: 0'][i])
len(categorical_features)

# top 20 and 40 feature lists 
features40 = [item for item in np.array(df_features40['Unnamed: 0'])]
features20 = [item for item in np.array(df_features40['Unnamed: 0'][:20])]

# remove columns that are not in our 40/20 features list
rm = []
features40.append('HERRFD')
features40.append('TOTRFD')
features20.append('HERRFD')
features20.append('TOTRFD')
for i in range(len(df_train_imp0.columns)):
    if df_train_imp0.columns[i] not in features40:
        rm.append(df_train_imp0.columns[i])
df_train40 = df_train_imp0.drop(columns=rm)
df_test40  = df_test_imp0.drop(columns=rm)

rm = []
for i in range(len(df_train_imp0.columns)):
    if df_train_imp0.columns[i] not in features20:
        rm.append(df_train_imp0.columns[i])
df_train20 = df_train_imp0.drop(columns=rm)
df_test20  = df_test_imp0.drop(columns=rm)

df_test20.head()

# get dummies
df_train40_encoded = pd.get_dummies(df_train40, columns=categorical_features)
df_test40_encoded  = pd.get_dummies(df_test40, columns=categorical_features)
df_train20_encoded = pd.get_dummies(df_train20, columns=categorical_features[:17])
df_test20_encoded  = pd.get_dummies(df_test20, columns=categorical_features[:17])

# Move RFD score columns to the end for convenience. 
cols_to_end = ['HERRFD','TOTRFD']
df_train40 = df_train40_encoded[[c for c in df_train40_encoded if c not in cols_to_end]+cols_to_end]
df_test40  = df_test40_encoded[[c for c in df_test40_encoded if c not in cols_to_end]+cols_to_end]

df_train20 = df_train20_encoded[[c for c in df_train20_encoded if c not in cols_to_end]+cols_to_end]
df_test20  = df_test20_encoded[[c for c in df_test20_encoded if c not in cols_to_end]+cols_to_end]

df_train40.head()

## Export

df_test40.to_csv('data/test40.tsv',sep='\t',index=False)
df_train40.to_csv('data/train40.tsv',sep='\t',index=False)
df_test20.to_csv('data/test20.tsv',sep='\t',index=False)
df_train20.to_csv('data/train20.tsv',sep='\t',index=False)
