==> 4_TrainAndEvaluate.ipynb <==
# Train and Evaluate Models

In this notebook XGBoost (logistic) Regressor, and a classic logit model are going to be trained and compared in terms performance against the unbalanced test data set. 

Varied for analysis:
* 20 vs 40 features
* XGBoost Regressor Model
* Heroin score vs accumulated score

import numpy as np
import pandas as pd

df_train40 = pd.read_csv("data/train40.tsv", sep="\t")
X_train40 = df_train40.iloc[:,0:143]
y_train40 = df_train40.iloc[:,143:145]

df_train20 = pd.read_csv("data/train20.tsv", sep="\t")
X_train20 = df_train20.iloc[:,0:63]
y_train20 = df_train20.iloc[:,63:65]

df_test40 = pd.read_csv("data/test40.tsv", sep="\t")
X_test40 = df_test40.iloc[:,0:143]
y_test40 = df_test40.iloc[:,143:145]

df_test20 = pd.read_csv("data/test20.tsv", sep="\t")
X_test20 = df_test20.iloc[:,0:63]
y_test20 = df_test20.iloc[:,63:65]

## Evaluation Functions

Define train & test function for each model

from xgboost import XGBRegressor
from sklearn.metrics import explained_variance_score, r2_score

def trainAndTestXGB(X_train, X_test, y_train, y_test):
    model = XGBRegressor(objective="reg:logistic")
    model.fit(X_train, y_train)
    
    y_pred = model.predict(X_test)
    r2 = r2_score(y_test, y_pred)
    print("R Square: ", r2)

## Analysis

# 40 features with Total RFD Score
trainAndTestXGB(X_train40, X_test40, y_train40.iloc[:,1], y_test40.iloc[:,1])

# 40 features with Heroin RFD Score
trainAndTestXGB(X_train40, X_test40, y_train40.iloc[:,0], y_test40.iloc[:,0])

# 20 features with Total RFD Score
trainAndTestXGB(X_train20, X_test20, y_train20.iloc[:,1], y_test20.iloc[:,1])

# 20 features with Heroin RFD Score
trainAndTestXGB(X_train20, X_test20, y_train20.iloc[:,0], y_test20.iloc[:,0])
